requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attr==0.3.1
attrs==24.2.0
azure-core==1.31.0
azure-storage-blob==12.23.1
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==5.0.1
boto==2.49.0
boto3==1.35.12
botocore==1.35.12
boxing==0.1.4
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
exceptiongroup==1.2.2
executing==2.1.0
expiringdict==1.2.2
fastapi==0.113.0
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.21.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.2
humansignal-drf-yasg==1.21.9
idna==3.8
ijson==3.3.0
importlib_metadata==8.4.0
inflection==0.5.1
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
label-studio==1.13.1
label-studio-sdk==1.0.5
launchdarkly-server-sdk==8.2.1
lightning==2.4.0
lightning-cloud==0.5.70
lightning-utilities==0.11.7
lightning_sdk==0.1.15
litdata==0.2.19
litserve==0.2.2
lockfile==0.12.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.52.2
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
ordered-set==4.0.2
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==4.23.4
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
pyRFC3339==1.1
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
python-multipart==0.0.9
pytorch-lightning==2.4.0
pytz==2022.7.1
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.38.0
redis==3.5.3
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.20.0
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
semver==3.0.2
Send2Trash==1.8.3
sentry-sdk==2.17.0
simple-term-menu==1.6.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.38.4
sympy==1.13.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
torch==2.2.1+cu121
torchmetrics==1.3.1
torchsummary==1.5.1
torchvision==0.17.1+cu121
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==2.2.0
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
ujson==5.10.0
ultralytics==8.3.23
ultralytics-thop==2.0.9
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
uvicorn==0.30.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
xmljson==0.2.1
yarl==1.9.11
zipp==3.20.2
